Добавялет общую логику обновления
